Move onvif base entity to separate module
