+++ homeassistant/components/onvif/entity.py
